refactored to use feature based folder structure as well as absolute imports
